pyproject.toml
[project]
name = "frida-freq"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.3",
]
